requirements.txt
Django==4.2.4
djangorestframework==3.15.2
aiogram==3.13.1
python-dotenv==1.0
requests==2.31.0
django-filter==24.3